requirements.txt
numpy
imutils
scikit-learn
tensorflow
keras
perlin-noise
flasgger
flask
opencv-python
Pillow
flask_healthz